requirements.txt
jiwer==3.1.0
matplotlib==3.7.5
numpy==1.24.4
pandas==2.0.3
python_Levenshtein==0.27.1
scikit_learn==1.3.2
seaborn==0.13.2
sentencepiece==0.2.0
tokenizers==0.21.1
torch==2.4.1+cu118
tqdm==4.67.0
transformers==4.46.1
wandb==0.19.11
